font is added in index.html, theme.jsx is created, modify on App.jsx (import are updated) / BottomBarStyles.jsx (marginbottom of icons)  and Map.jsx (vheight =>85vh). App.scss is deleted
