+++ src/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
+import { MuiThemeProvider } from '@material-ui/core/styles';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import BottomBar from './components/bottombar/BottomBar';
 import Explore from './components/explore/Explore';
@@ -8,31 +8,8 @@ import Leaflet from './components/map/Map';
 import Parcours from './components/parcours/Parcours';
 import Profil from './components/profil/Profil';
 import NavBar from './components/navbar/Navbar';
-import './App.scss';
 import 'leaflet/dist/leaflet.css';
-
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      main: '#F15348',
-    },
-    secondary: {
-      main: '#FFFFFF',
-    },
-  },
-  typography: {
-    fontFamily: [
-      '-apple-system',
-      'BlinkMacSystemFont',
-      '"Montserrat"',
-      'Arial',
-      'sans-serif',
-      '"Apple Color Emoji"',
-      '"Segoe UI Emoji"',
-      '"Segoe UI Symbol"',
-    ].join(','),
-  },
-});
+import theme from './Theme';
 
 function App() {
   return (
