+++ src/components/map/Map.jsx
@@ -5,7 +5,7 @@ const Leaflet = () => {
   const position = [50.63333, 3.066667];
 
   return (
-    <Map center={position} zoom={22} style={{ height: '100vh', zIndex: 0 }}>
+    <Map center={position} zoom={22} style={{ height: '85vh', zIndex: 0 }}>
       <TileLayer
         url="https://{s}.tile.openstreetmap.fr/osmfr/{z}/{x}/{y}.png"
         attribution='&copy; Openstreetmap France | &copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
